+++ NEWS.md
@@ -1,6 +1,8 @@
 ## v 0.1.0.99
 
 * install selected packages in gadget table
+    * with / without dependencies (pkgs in `Suggests` field)
+* add refresh button to reload the app
 
 ## upnews 0.1